+++ src/main/java/com/realworld/project/application/article/repository/TagRepository.java
@@ -1,5 +1,6 @@
-package com.realworld.project.application.article.domain;
+package com.realworld.project.application.article.repository;
 
+import com.realworld.project.application.article.domain.Tag;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TagRepository extends JpaRepository<Tag, Long> {
